% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppGE_back_elim.R
\name{mppGE_back_elim}
\alias{mppGE_back_elim}
\title{Backward elimination MPP GxE analysis}
\usage{
mppGE_back_elim(mppData, trait, Q.eff = "cr", VCOV = "CS_CSRT",
  QTL = NULL, alpha = 0.01, workspace = 8e+06)
}
\arguments{
\item{mppData}{An object of class \code{mppData}.}

\item{trait}{\code{Character vector} specifying which traits (environments) should be used.}

\item{Q.eff}{\code{Character} expression indicating the assumption concerning
the QTL effects: 1) "cr" for cross-specific; 2) "par" for parental; 3) "anc"
for ancestral; 4) "biall" for a bi-allelic. Default = "cr".}

\item{VCOV}{VCOV \code{Character} expression defining the type of variance
covariance structure used. "ID" for identity, "CSRT" for within environment
cross-specific residual terms, "CS_CSRT" for compound symmetry with within
environment cross-specific residual terms. Default = "CS_CSRT".}

\item{QTL}{Object of class \code{QTLlist} representing a list of
selected marker positions obtained with the function QTL_select() or
a vector of \code{character} marker positions names. Default = NULL.}

\item{alpha}{\code{Numeric} value indicating the level of significance for
the backward elimination. Default = 0.01.}

\item{workspace}{Size of workspace for the REML routines measured in double
precision words (groups of 8 bytes). The default is workspace = 8e6.}
}
\value{
Return:

\item{QTL }{\code{Data.frame} of class \code{QTLlist} with five columns :
1) QTL marker names; 2) chromosomes;
3) interger position indicators on the chromosome;
4) positions in centi-Morgan; and 5) -log10(p-values).}
}
\description{
Perform a backward elimination on a list of QTL using a MPP GxE model.
}
\examples{

library(asreml)

data(mppData_GE)

Qpos <- c("PZE.105068880", "PZE.106098900")

QTL <- mppGE_back_elim(mppData = mppData_GE, trait = c('DMY_CIAM', 'DMY_TUM'),
                       Q.eff = 'par', QTL = Qpos)

}
\author{
Vincent Garin
}
