% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppGE_CIM.R
\name{mppGE_CIM}
\alias{mppGE_CIM}
\title{MPP GxE composite interval maping}
\usage{
mppGE_CIM(
  mppData,
  trait,
  Q.eff = "cr",
  VCOV = "CS_CSRT",
  cofactors = NULL,
  window = 20,
  plot.gen.eff = FALSE,
  parallel = FALSE,
  cluster = NULL,
  workspace = 8e+06
)
}
\arguments{
\item{mppData}{An object of class \code{mppData}.}

\item{trait}{\code{Character vector} specifying which traits (environments) should be used.}

\item{Q.eff}{\code{Character} expression indicating the assumption concerning
the QTL effects: 1) "cr" for cross-specific; 2) "par" for parental; 3) "anc"
for ancestral; 4) "biall" for a bi-allelic. Default = "cr".}

\item{VCOV}{VCOV \code{Character} expression defining the type of variance
covariance structure used. "ID" for identity, "CSRT" for within environment
cross-specific residual terms, "CS_CSRT" for compound symmetry with within
environment cross-specific residual terms. Default = "CS_CSRT".}

\item{cofactors}{Object of class \code{QTLlist} representing a list of
selected marker positions obtained with the function QTL_select() or
a vector of \code{character} marker positions names.
Default = NULL.}

\item{window}{\code{Numeric} distance (cM) on the left and the right of a
cofactor position where it is not included in the model. Default = 20.}

\item{plot.gen.eff}{\code{Logical} value. If \code{plot.gen.eff = TRUE},
the function will save the significance of the QTL allelic effects per
cross/parent along the genome. These results can be visualized with the
function \code{\link{plot_genEffects_GE}}. Default value = FALSE.}

\item{parallel}{\code{Logical} value specifying if the function should be
executed in parallel on multiple cores. To run the function in parallel user must
provide clusters in the \code{cluster} argument. \strong{Parallelization is
only available for 'ID' model}. Default = FALSE.}

\item{cluster}{Cluster object obtained with the function \code{makeCluster()}
from the parallel package. Default = NULL.}

\item{workspace}{Size of workspace for the REML routines measured in double
precision words (groups of 8 bytes). The default is workspace = 8e6.}
}
\value{
Return:

\item{CIM }{\code{Data.frame} of class \code{QTLprof}. with five columns :
1) QTL marker or in between position names; 2) chromosomes;
3) interger position indicators on the chromosome;
4) positions in centi-Morgan; and 5) -log10(p-val). And if
\code{plot.gen.eff = TRUE}, p-values of the cross or parental QTL allelic effects.}
}
\description{
Compute QTL models along the genome using cofactors representing other
genetic positions for control.
}
\examples{

library(asreml)
library(mppR)

data(mppData_GE)

SIM <- mppGE_SIM(mppData = mppData_GE, trait = c('DMY_CIAM', 'DMY_TUM'),
                 Q.eff = 'par', plot.gen.eff = TRUE)

cofactors <- QTL_select(Qprof = SIM, threshold = 3, window = 50)

CIM <- mppGE_CIM(mppData = mppData_GE, trait = c('DMY_CIAM', 'DMY_TUM'),
                 Q.eff = 'par', cofactors = cofactors, plot.gen.eff = TRUE)

Qpos <- QTL_select(Qprof = CIM, threshold = 3, window = 20)

plot(x = CIM, QTL = Qpos)

plot_genEffects_GE(mppData = mppData_GE, nEnv = 2, EnvNames = c('CIAM', 'TUM'),
                   Qprof = CIM, Q.eff = 'par', QTL = Qpos, text.size = 14)

}
\seealso{
\code{\link{mppGE_SIM}}, \code{\link{plot_genEffects_GE}}
}
\author{
Vincent Garin
}
