% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppGE_CIM_fast.R
\name{mppGE_CIM_fast}
\alias{mppGE_CIM_fast}
\title{MPP GxE Composite Interval Mapping using fast MM GLS algorithm}
\usage{
mppGE_CIM_fast(
  mppData,
  trait,
  Q.eff = "cr",
  VCOV = "UN",
  cofactors = NULL,
  window = 20,
  n.cores = 1,
  maxIter = 100,
  msMaxIter = 100
)
}
\arguments{
\item{mppData}{An object of class \code{mppData}.}

\item{trait}{\code{Character vector} specifying which traits (environments) should be used.}

\item{Q.eff}{\code{Character} expression indicating the assumption concerning
the QTL effects: 1) "cr" for cross-specific; 2) "par" for parental; 3) "anc"
for ancestral; 4) "biall" for a bi-allelic. Default = "cr".}

\item{VCOV}{VCOV \code{Character} expression defining the type of variance
covariance structure used. 'CS' for compound symmetry assuming a unique
genetic covariance between environments. 'CSE' for cross-specific within
environment error term. 'CS_CSE' for both compound symmetry plus
cross-specific within environment error term. 'UN' for unstructured
environmental variance covariance structure allowing a specific genotypic
covariance for each pair of environments. Default = 'UN'}

\item{cofactors}{Object of class \code{QTLlist} representing a list of
selected marker positions obtained with the function QTL_select() or
a vector of \code{character} marker positions names.
Default = NULL.}

\item{window}{\code{Numeric} distance (cM) on the left and the right of a
cofactor position where it is not included in the model. Default = 20.}

\item{n.cores}{\code{Numeric}. Specify here the number of cores you like to
use. Default = 1.}

\item{maxIter}{maximum number of iterations for the lme optimization algorithm.
Default = 100.}

\item{msMaxIter}{maximum number of iterations for the optimization step inside
the lme optimization. Default = 100.}
}
\value{
Return:

\item{CIM }{\code{Data.frame} of class \code{QTLprof}. with five columns :
1) QTL marker or in between position names; 2) chromosomes;
3) interger position indicators on the chromosome;
4) positions in centi-Morgan; 5) -log10(p-val); and 6 onwards) if
p-values of the cross or parental QTL allelic effects.}
}
\description{
Computes a multi QTL models along the genome using different models. Add extra
elements to provide more information about the procedure. The initial mixed
are calculated with functions from the nlme package.
}
\examples{

library(mppR)

data(mppData_GE)

cofactors <- mppData_GE$map$mk.names[c(35, 61)]

CIM <- mppGE_CIM_fast(mppData = mppData_GE, trait = c('DMY_CIAM', 'DMY_TUM'),
                 Q.eff = 'par')

Qpos <- QTL_select(Qprof = CIM, threshold = 3, window = 50)

plot.QTLprof(x = CIM, QTL = Qpos)

plot_genEffects_GE(mppData = mppData_GE, nEnv = 2, EnvNames = c('CIAM', 'TUM'),
                   Qprof = CIM, Q.eff = 'par', QTL = Qpos, text.size = 14)

}
\references{
Pinheiro J, Bates D, DebRoy S, Sarkar D, R Core Team (2021). nlme: Linear
and Nonlinear Mixed Effects Models_. R package version 3.1-152,
<URL: https://CRAN.R-project.org/package=nlme>.
}
\seealso{
\code{\link{mppGE_SIM_fast}}
}
\author{
Vincent Garin
}
