% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppGE_CV.R
\name{mppGE_CV}
\alias{mppGE_CV}
\title{MPP GxE cross-validation}
\usage{
mppGE_CV(pop.name = "MPP", trait.name = "trait1", mppData, trait,
  cv.ref = NULL, Rep = 5, k = 3, EnvNames = NULL, Q.eff = "cr",
  VCOV = "CS_CSRT", thre.cof = 4, win.cof = 50, N.cim = 1,
  window = 20, thre.QTL = 4, win.QTL = 20, alpha = 0.01,
  parallel = FALSE, cluster = NULL, workspace = 8e+06, verbose = TRUE,
  output.loc = NULL)
}
\arguments{
\item{pop.name}{\code{Character} name of the studied population.
Default = "MPP".}

\item{trait.name}{\code{Character} name of the studied trait.
Default = "trait1".}

\item{mppData}{An object of class \code{mppData}.}

\item{trait}{\code{Character vector} specifying which traits (environments) should be used.}

\item{cv.ref}{\code{Numerical} or \code{character} indicator to specify which
trait of the \code{mppData} object should be used to check the prediction
in the CV process. By default use 'trait'.}

\item{Rep}{\code{Numeric} value representing the number of repetitions of the
k-fold procedure. Default = 5.}

\item{k}{\code{Numeric} value representing the number of folds for the within
cross partition of the population. Default = 3.}

\item{EnvNames}{\code{Character} expression indicating the environment or trait
labels. By default: Env_1, 2, 3, etc.}

\item{Q.eff}{\code{Character} expression indicating the assumption concerning
the QTL effect: 1) "cr" for cross-specific effects; 2) "par" parental
effects; 3) "anc" for an ancestral effects; 4) "biall" for a bi-allelic
effects. Default = "cr".}

\item{VCOV}{VCOV \code{Character} expression defining the type of variance
covariance structure used. "ID" for identity, "CSRT" for within environment
cross-specific residual terms, "CS_CSRT" for compound symmetry with within
environment cross-specific residual terms. Default = "CS_CSRT".}

\item{thre.cof}{\code{Numeric} value representing the -log10(p-value)
threshold above which a position can be selected as cofactor. Default = 4.}

\item{win.cof}{\code{Numeric} value in centi-Morgan representing the minimum
distance between two selected cofactors. Default = 50 cM.}

\item{N.cim}{\code{Numeric} value specifying the number of times the CIM
analysis is repeated. Default = 1.}

\item{window}{\code{Numeric} distance (cM) on the left and the right of a
cofactor position where it is not included in the model. Default = 20.}

\item{thre.QTL}{\code{Numeric} value representing the -log10(p-value)
threshold above which a position can be selected as QTL. Default = 4.}

\item{win.QTL}{\code{Numeric} value in centi-Morgan representing the minimum
distance between two selected QTLs. Default = 20.}

\item{alpha}{\code{Numeric} value indicating the level of significance for
the backward elimination. Default = 0.01.}

\item{parallel}{\code{Logical} value specifying if the function should be
executed in parallel on multiple cores. To run the function in parallel user must
provide clusters in the \code{cluster} argument. \strong{Parallelization is
only available for 'ID' model}. Default = FALSE.}

\item{cluster}{Cluster object obtained with the function \code{makeCluster()}
from the parallel package. Default = NULL.}

\item{workspace}{Size of workspace for the REML routines measured in double
precision words (groups of 8 bytes). The default is workspace = 8e6.}

\item{verbose}{\code{Logical} value indicating if the steps of mpp_proc should
be printed. It will not affect the printing of the other functions called by
\code{mpp_proc()}, especially the printing of \code{asreml()}. Default = TRUE.}

\item{output.loc}{Path where a folder will be created to save the results.
Default = NULL.}
}
\value{
Return:

\code{List} containing the following results items:

\item{p_vs}{\code{Matrix} with : 1) the number of detected QTL;
2) the proportion of predicted genetic variance
in the VS (p.vs) at the population level (average of within cross predictions)
per environment.}

\item{QTL}{\code{Data.frame} containing: 1) the list of QTL position detected
at least one time during the entire CV process; 2) the number of times
the position has been detected}

The results elements returned as R object are also saved as text
files at the specified output location (\code{output.loc}).
}
\description{
MPP GxE cross-validation.
}
\examples{

\dontrun{

library(asreml)

data(mppData_GE)

# Specify a location where your results will be saved
my.loc <- tempdir()

QTL <- mppGE_CV(pop.name = 'EUNAM', trait.name = 'DMY', mppData = mppData_GE,
                trait = c('DMY_CIAM', 'DMY_TUM'), Rep = 1, k = 2,
                EnvNames = c('CIAM', 'TUM'),  Q.eff = 'par',
                thre.cof = 3, thre.QTL = 3, verbose = FALSE,
                output.loc = my.loc)

}

}
\author{
Vincent Garin
}
