% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppGE_oneS_back_elim.R
\name{mppGE_oneS_back_elim}
\alias{mppGE_oneS_back_elim}
\title{Backward elimination of one stage MPP GxE analysis}
\usage{
mppGE_oneS_back_elim(plot_data, mppData, trait, Q.eff = "cr",
  VCOV = "CS_CSRT", exp_des_form, QTL = NULL, alpha = 0.01,
  workspace = 8e+06)
}
\arguments{
\item{plot_data}{\code{Data.frame} containing the plot data with the following
columns: the trait(s), 'genotype' (genotype indicator), 'check'
(check indicator), 'cross' (cross indicator), 'env' (environment indicator),
and all other experimental design covariates (e.g. replicate, blocks, etc.).
The column names must be ('genotype', 'check', 'cross', env'). The names of
the experimental design covariates must be the same as the one used in
'exp_des_form'. For more details see \code{\link{plot_data}}.}

\item{mppData}{Object of class \code{mppData} contaning the the same
genotype identifiers as the one in \code{plot_data} ('genotype').}

\item{trait}{\code{Character} expression for the trait matching the trait
column in 'plot_data' argument.}

\item{Q.eff}{\code{Character} expression indicating the assumption concerning
the QTL effects: 1) "cr" for cross-specific; 2) "par" for parental; 3) "anc"
for ancestral; 4) "biall" for a bi-allelic. Default = "cr".}

\item{VCOV}{VCOV \code{Character} expression defining the type of variance
covariance structure used: a) "CSRT" for within environment
cross-specific residual terms; b) "CS_CSRT" for compound symmetry with within
environment cross-specific residual terms. Default = "CS_CSRT".}

\item{exp_des_form}{\code{Character} expression for the random experimental
design effects in asreml-R format. For example,
'env:replicate + env:replicate:block'. The variable names used in
'exp_des_form' should strictly match the column names used in 'plot_data'.}

\item{QTL}{Object of class \code{QTLlist} representing a list of
selected marker positions obtained with the function QTL_select() or
a vector of \code{character} marker positions names. Default = NULL.}

\item{alpha}{\code{Numeric} value indicating the level of significance for
the backward elimination. Default = 0.01.}

\item{workspace}{Size of workspace for the REML routines measured in double
precision words (groups of 8 bytes). The default is workspace = 8e6.}
}
\value{
Return:

\item{QTL }{\code{Data.frame} of class \code{QTLlist} with five columns :
1) QTL marker names; 2) chromosomes;
3) interger position indicators on the chromosome;
4) positions in centi-Morgan; and 5) -log10(p-values).}
}
\description{
Backward elimination on a list of QTL using a one stage MPP GxE
model.
}
\examples{

library(asreml)

data(mppData_GE)
data(plot_data)

Qpos <- c("PZE.105068880", "PZE.106098900")

QTL <- mppGE_oneS_back_elim(plot_data = plot_data, mppData = mppData_GE,
                            trait = 'DMY', Q.eff = 'par',
                            exp_des_form = 'env:Rep + env:Rep:Block',
                            QTL = Qpos)

}
\author{
Vincent Garin
}
