% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppGE_oneS_CIM.R
\name{mppGE_oneS_CIM}
\alias{mppGE_oneS_CIM}
\title{MPP GxE one stage CIM analysis}
\usage{
mppGE_oneS_CIM(
  plot_data,
  mppData,
  trait,
  Q.eff = "cr",
  VCOV = "CS_CSRT",
  exp_des_form,
  cofactors = NULL,
  window = 20,
  plot.gen.eff = FALSE,
  workspace = 8e+06
)
}
\arguments{
\item{plot_data}{\code{Data.frame} containing the plot data with the following
columns: the trait(s), 'genotype' (genotype indicator), 'check'
(check indicator), 'cross' (cross indicator), 'env' (environment indicator),
and all other experimental design covariates (e.g. replicate, blocks, etc.).
The column names must be ('genotype', 'check', 'cross', env'). The names of
the experimental design covariates must be the same as the one used in
'exp_des_form'. For more details see \code{\link{plot_data}}.}

\item{mppData}{Object of class \code{mppData} contaning the the same
genotype identifiers as the one in \code{plot_data} ('genotype').}

\item{trait}{\code{Character} expression for the trait matching the trait
column in 'plot_data' argument.}

\item{Q.eff}{\code{Character} expression indicating the assumption concerning
the QTL effects: 1) "cr" for cross-specific; 2) "par" for parental; 3) "anc"
for ancestral; 4) "biall" for a bi-allelic. Default = "cr".}

\item{VCOV}{VCOV \code{Character} expression defining the type of variance
covariance structure used: a) "CSRT" for within environment
cross-specific residual terms; b) "CS_CSRT" for compound symmetry with within
environment cross-specific residual terms. Default = "CS_CSRT".}

\item{exp_des_form}{\code{Character} expression for the random experimental
design effects in asreml-R format. For example,
'env:replicate + env:replicate:block'. The variable names used in
'exp_des_form' should strictly match the column names used in 'plot_data'.}

\item{cofactors}{Object of class \code{QTLlist} representing a list of
selected marker positions obtained with the function QTL_select() or
a vector of \code{character} marker positions names.
Default = NULL.}

\item{window}{\code{Numeric} distance (cM) on the left and the right of a
cofactor position where it is not included in the model. Default = 20.}

\item{plot.gen.eff}{\code{Logical} value. If \code{plot.gen.eff = TRUE},
the function will save the significance of the QTL allelic effects per
cross/parent along the genome. These results can be visualized with the
function \code{\link{plot_genEffects_GE}}. Default value = FALSE.}

\item{workspace}{Size of workspace for the REML routines measured in double
precision words (groups of 8 bytes). The default is workspace = 8e6.}
}
\value{
Return:

\item{CIM }{\code{Data.frame} of class \code{QTLprof}. with five columns :
1) QTL marker or in between position names; 2) chromosomes;
3) interger position indicators on the chromosome;
4) positions in centi-Morgan; and 5) -log10(p-val). And if
\code{plot.gen.eff = TRUE}, p-values of the cross or parental QTL allelic effects.}
}
\description{
Perform a MPP GxE one stage CIM analysis.
}
\examples{

\dontrun{

library(asreml)

data(mppData_GE)
data(plot_data)


cofactors <- c("PZE.105068880", "PZE.106098900")

# takes 10 min

CIM <- mppGE_oneS_CIM(plot_data = plot_data, mppData = mppData_GE,
                      trait = 'DMY', Q.eff = 'par',
                      exp_des_form = 'env:Rep + env:Rep:Block',
                      cofactors = cofactors, plot.gen.eff = TRUE)

plot(x = CIM)

plot_genEffects_GE(mppData = mppData_GE, nEnv = 2, EnvNames = c('CIAM', 'TUM'),
                   Qprof = CIM, Q.eff = 'par', QTL = Qpos, text.size = 14)


}

}
\seealso{
\code{\link{plot_genEffects_GE}}
}
\author{
Vincent Garin
}
