% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppGE_oneS_CV.R
\name{mppGE_oneS_CV}
\alias{mppGE_oneS_CV}
\title{MPP GxE cross-validation one stage analysis}
\usage{
mppGE_oneS_CV(pop.name = "MPP", trait.name = "trait1", plot_data, mppData,
  trait, cv.ref, Rep = 5, k = 3, EnvNames = NULL, Q.eff = "cr",
  VCOV = "CS_CSRT", exp_des_form, thre.cof = 4, win.cof = 50, N.cim = 1,
  window = 20, thre.QTL = 4, win.QTL = 20, alpha = 0.01,
  verbose = TRUE, output.loc = NULL, workspace = 8e+06)
}
\arguments{
\item{pop.name}{\code{Character} name of the studied population.
Default = "MPP".}

\item{trait.name}{\code{Character} name of the studied trait.
Default = "trait1".}

\item{plot_data}{\code{Data.frame} containing the plot data with the following
columns: the trait(s), 'genotype' (genotype indicator), 'check'
(check indicator), 'cross' (cross indicator), 'env' (environment indicator),
and all other experimental design covariates (e.g. replicate, blocks, etc.).
The column names must be ('genotype', 'check', 'cross', env'). The names of
the experimental design covariates must be the same as the one used in
'exp_des_form'. For more details see \code{\link{plot_data}}.}

\item{mppData}{Object of class \code{mppData} contaning the the same
genotype identifiers as the one in \code{plot_data} ('genotype').}

\item{trait}{\code{Character} expression for the trait matching the trait
column in 'plot_data' argument.}

\item{cv.ref}{\code{Character} indicator to specify which
trait of the \code{mppData} object should be used to evaluate the prediction
accuracy in the CV process. Possibility to specify a vector to evaluate
prediction with respect to different within environment genotypic vector
values.}

\item{Rep}{\code{Numeric} value representing the number of repetitions of the
k-fold procedure. Default = 5.}

\item{k}{\code{Numeric} value representing the number of folds for the within
cross partition of the population. Default = 3.}

\item{EnvNames}{\code{character} expression indicating the environment or trait
labels. By default: Env_1, 2, 3, etc.}

\item{Q.eff}{\code{Character} expression indicating the assumption concerning
the QTL effect: 1) "cr" for cross-specific effects; 2) "par" parental
effects; 3) "anc" for an ancestral effects; 4) "biall" for a bi-allelic
effects. Default = "cr".}

\item{VCOV}{VCOV \code{Character} expression defining the type of variance
covariance structure used: a) "CSRT" for within environment
cross-specific residual terms; b) "CS_CSRT" for compound symmetry with within
environment cross-specific residual terms. Default = "CS_CSRT".}

\item{exp_des_form}{\code{Character} expression for the random experimental
design effects in asreml-R format. For example,
'env:replicate + env:replicate:block'. The variable names used in
'exp_des_form' should strictly match the column names used in 'plot_data'.}

\item{thre.cof}{\code{Numeric} value representing the -log10(p-value)
threshold above which a position can be selected as cofactor. Default = 4.}

\item{win.cof}{\code{Numeric} value in centi-Morgan representing the minimum
distance between two selected cofactors. Default = 50 cM.}

\item{N.cim}{\code{Numeric} value specifying the number of times the CIM
analysis is repeated. Default = 1.}

\item{window}{\code{Numeric} distance (cM) on the left and the right of a
cofactor position where it is not included in the model. Default = 20.}

\item{thre.QTL}{\code{Numeric} value representing the -log10(p-value)
threshold above which a position can be selected as QTL. Default = 4.}

\item{win.QTL}{\code{Numeric} value in centi-Morgan representing the minimum
distance between two selected QTLs. Default = 20.}

\item{alpha}{\code{Numeric} value indicating the level of significance for
the backward elimination. Default = 0.01.}

\item{verbose}{\code{Logical} value indicating if the steps of mpp_proc should
be printed. It will not affect the printing of the other functions called by
\code{mpp_proc()}, especially the printing of \code{asreml()}. Default = TRUE.}

\item{output.loc}{Path where a folder will be created to save the results.
Default = NULL.}

\item{workspace}{Size of workspace for the REML routines measured in double
precision words (groups of 8 bytes). The default is workspace = 8e6.}
}
\value{
Return:

\code{List} containing the following results items:

\item{p_vs}{\code{Matrix} with : 1) the number of detected QTL;
2) the proportion of predicted genetic variance
in the VS (p.vs) at the population level (average of within cross predictions)
per environment.}

\item{QTL}{\code{Data.frame} containing: 1) the list of QTL position detected
at least one time during the entire CV process; 2) the number of times
the position has been detected}

The results elements returned as R object are also saved as text
files at the specified output location (\code{output.loc}).
}
\description{
MPP GxE cross-validation one stage analysis
}
\examples{

\dontrun{

library(asreml)

data(mppData_GE)
data(plot_data)

# Specify a location where your results will be saved
my.loc <- tempdir()

CV <- mppGE_oneS_CV(pop.name = 'EUNAM', trait.name = 'DMY',
                    plot_data = plot_data, mppData = mppData_GE,
                    trait = 'DMY', cv.ref = c('DMY_CIAM', 'DMY_TUM'),
                    Rep = 1, k = 2, EnvNames = c('CIAM', 'TUM'),
                    Q.eff = 'par', exp_des_form = 'env:Rep + env:Rep:Block',
                    thre.cof = 3,thre.QTL = 3, verbose = FALSE,
                    output.loc = my.loc)

}

}
\author{
Vincent Garin
}
