% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppGE_oneS_proc.R
\name{mppGE_oneS_proc}
\alias{mppGE_oneS_proc}
\title{MPP GxE one stage QTL analysis}
\usage{
mppGE_oneS_proc(pop.name = "MPP", trait.name = "trait1", plot_data, mppData,
  trait, EnvNames = NULL, Q.eff = "cr", VCOV = "CS_CSRT", exp_des_form,
  plot.gen.eff = FALSE, thre.cof = 4, win.cof = 50, N.cim = 1,
  window = 20, thre.QTL = 4, win.QTL = 20, alpha = 0.01,
  Qeff_est = TRUE, text.size = 18, verbose = TRUE, output.loc = NULL,
  workspace = 8e+06)
}
\arguments{
\item{pop.name}{\code{Character} name of the studied population.
Default = "MPP".}

\item{trait.name}{\code{Character} name of the studied trait.
Default = "trait1".}

\item{plot_data}{\code{data.frame} containing the plot data with the following
columns: the trait(s), 'genotype' (genotype indicator), 'check'
(check indicator), 'cross' (cross indicator), 'env' (environment indicator),
and all other experimental design covariates (e.g. replicate, blocks, etc.).
The column names of the data.frame must be identical to the one specified
('genotype', 'check', 'cross', env'). The names of the experimental design
covariates must be the same as the one used in 'exp_des_form'. for more
details see \code{\link{plot_data}}.}

\item{mppData}{Object of class \code{mppData} contaning the genotypic
information with genotype list corresponding to the one of \code{plot_data}.}

\item{trait}{\code{Character} expression for the trait matching the trait
column in 'plot_data' argument.}

\item{EnvNames}{\code{character} expression indicating the environment or trait
labels.}

\item{Q.eff}{\code{Character} expression indicating the assumption concerning
the QTL effect: 1) "cr" for cross-specific effects; 2) "par" parental
effects; 3) "anc" for an ancestral effects; 4) "biall" for a bi-allelic
effects. Default = "cr".}

\item{VCOV}{VCOV \code{Character} expression defining the type of variance
covariance structure used: a) "CSRT" for within environment
cross-specific residual term; b) "CS_CSRT" for compound symmetry with within
environment cross-specific residual term. Default = "CS_CSRT".}

\item{exp_des_form}{\code{Character} expression for the random experimental
design effects in asreml-R format. For example,
'env:replicate + env:replicate:block'. The column variables names used in
'exp_des_form' should strictly match the names used in 'plot_data'.}

\item{plot.gen.eff}{\code{Logical} value. If \code{plot.gen.eff = TRUE},
the function will save the decomposed genetic effects per cross/parent.
These results can be ploted with the function \code{\link{plot_genEffects_GE}}
to visualize a genome-wide decomposition of the genetic effects.
Default value = FALSE.}

\item{thre.cof}{\code{Numeric} value representing the -log10(p-value)
threshold above which a position can be peaked as a cofactor. Default = 4.}

\item{win.cof}{\code{Numeric} value in centi-Morgan representing the minimum
distance between two selected cofactors. Default = 50 cM.}

\item{N.cim}{\code{Numeric} value specifying the number of time the CIM
analysis is repeated. Default = 1.}

\item{window}{\code{Numeric} distance (cM) on the left and the right of a
cofactor position where it is not included in the model. Default = 20.}

\item{thre.QTL}{\code{Numeric} value representing the -log10(p-value)
threshold above which a position can be selected as QTL. Default = 4.}

\item{win.QTL}{\code{Numeric} value in centi-Morgan representing the minimum
distance between two selected QTLs. Default = 20.}

\item{alpha}{\code{Numeric} value indicating the level of significance for
the backward elimination. Default = 0.01.}

\item{Qeff_est}{\code{Logical} value specifying if the QTL effects of the
final QTL list should be estimated. Default = TRUE.}

\item{text.size}{\code{Numeric} value specifying the size of graph axis text
elements. Default = 18.}

\item{verbose}{\code{Logical} value indicating if the steps of mpp_proc should
be printed. It will not affect the printing of the other functions called by
\code{mpp_proc()}, especially the printing of \code{asreml()}. Default = TRUE.}

\item{output.loc}{Path where a folder will be created to save the results.
Default = NULL.}

\item{workspace}{size of workspace for the REML routines measured in double
precision words (groups of 8 bytes). The default is workspace = 8e6.}
}
\value{
Return:

List containing the following items:

\item{n.QTL}{Number of detected QTLs.}

\item{cofactors}{\code{Data.frame} with cofactors positions.}

\item{QTL}{\code{Data.frame} with QTL positions.}

Some output files are also saved at the specified location
(\code{output.loc}):

\enumerate{

\item{The SIM and CIM results in a text file (SIM.txt, CIM.txt).}

\item{The list of cofactors (cofactors.txt).}

\item{The list of QTL (QTL.txt).}

\item{The plot of the CIM profile (QTL_profile.pdf) with dotted vertical
lines representing the cofactors positions. If \code{plot.gen.eff = TRUE},
plot of the genetic effects per cross or parents obtained with
\code{\link{plot_genEffects_GE}}  (gen_eff.pdf) with dashed
lines representing the QTL positions.}

}
}
\description{
The procedure is the following:
}
\details{
\enumerate{

\item{Simple interval mapping (SIM) to select cofactor
(\code{\link{mppGE_oneS_SIM}}).}

\item{Composite interval mapping (CIM) with selected cofactors
(\code{\link{mppGE_oneS_CIM}}).}

\item{Backward elimination on the list of QTLs
(\code{\link{mppGE_oneS_back_elim}}).}

}
}
\examples{

\dontrun{

library(asreml)

data(mppData_GE)
data(plot_data)

# Specify a location where your results will be saved
my.loc <- tempdir()

QTL <- mppGE_oneS_proc(pop.name = 'EUNAM', trait.name = 'DMY',
                       plot_data = plot_data, mppData = mppData_GE,
                       trait = 'DMY', EnvNames = c('CIAM', 'TUM'),
                       Q.eff = 'par', exp_des_form = 'env:Rep + env:Rep:Block',
                       plot.gen.eff = TRUE, thre.cof = 3,thre.QTL = 3,
                       verbose = FALSE, output.loc = my.loc)

}


}
\seealso{
\code{\link{mppGE_CIM}},
\code{\link{mppGE_SIM}},
\code{\link{plot_genEffects_GE}},
}
\author{
Vincent Garin
}
