% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppGE_oneS_QTL_effects.R
\name{mppGE_oneS_QTL_effects}
\alias{mppGE_oneS_QTL_effects}
\title{MPP GxE one stage QTL genetic effects}
\usage{
mppGE_oneS_QTL_effects(plot_data, mppData, trait, Q.eff = "cr",
  VCOV = "CS_CSRT", exp_des_form, QTL = NULL, workspace = 8e+06)
}
\arguments{
\item{plot_data}{\code{data.frame} containing the plot data with the following
columns: the trait(s), 'genotype' (genotype indicator), 'check'
(check indicator), 'cross' (cross indicator), 'env' (environment indicator),
and all other experimental design covariates (e.g. replicate, blocks, etc.).
The column names of the data.frame must be identical to the one specified
('genotype', 'check', 'cross', env'). The names of the experimental design
covariates must be the same as the one used in 'exp_des_form'. for more
details see \code{\link{plot_data}}.}

\item{mppData}{An object of class \code{mppData}.}

\item{trait}{\code{Character} expression for the trait matching the trait
column in 'plot_data' argument.}

\item{Q.eff}{\code{Character} expression indicating the assumption concerning
the QTL effects: 1) "cr" for cross-specific; 2) "par" for parental; 3) "anc"
for ancestral; 4) "biall" for a bi-allelic. Default = "cr".}

\item{VCOV}{VCOV \code{Character} expression defining the type of variance
covariance structure used: a) "CSRT" for within environment
cross-specific residual term; b) "CS_CSRT" for compound symmetry with within
environment cross-specific residual term. Default = "CS_CSRT".}

\item{exp_des_form}{\code{Character} expression for the random experimental
design effects in asreml-R format. For example,
'env:replicate + env:replicate:block'. The column variables names used in
'exp_des_form' should strictly match the names used in 'plot_data'.}

\item{QTL}{Object of class \code{QTLlist} representing a list of
selected marker positions obtained with the function QTL_select() or
vector of \code{character} marker positions names. Default = NULL.}

\item{workspace}{size of workspace for the REML routines measured in double
precision words (groups of 8 bytes). The default is workspace = 8e6.}
}
\value{
Return:

\item{Qeff}{\code{List} of \code{data.frame} (one per QTL) containing the
following information:

\enumerate{

\item{QTL genetic effects}
\item{Standard error of the QTL effects.}
\item{Test statistics of the effects (t-test or Wald statistic).}
\item{P-value of the test statistics.}
\item{Significance of the QTL effects.}

}

}
}
\description{
Compute MPP GxE one stage QTL genetic effects
}
\examples{

library(asreml)

data(mppData_GE)
data(plot_data)

Qpos <- c("PZE.105068880", "PZE.106098900")

Qeff <- mppGE_oneS_QTL_effects(plot_data = plot_data, mppData = mppData_GE,
                              trait = 'DMY', Q.eff = 'par',
                              exp_des_form = 'env:Rep + env:Rep:Block',
                              QTL = Qpos)

Qeff

}
\author{
Vincent Garin
}
