% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppGE_oneS_QTL_R2.R
\name{mppGE_oneS_QTL_R2}
\alias{mppGE_oneS_QTL_R2}
\title{MPP GxE one stage QTL R2}
\usage{
mppGE_oneS_QTL_R2(plot_data, mppData, trait, Q.eff = "cr", VCOV = "CS_CSRT",
  exp_des_form, QTL = NULL, glb.only = FALSE, workspace = 8e+06)
}
\arguments{
\item{plot_data}{\code{data.frame} containing the plot data with the following
columns: the trait(s), 'genotype' (genotype indicator), 'check'
(check indicator), 'cross' (cross indicator), 'env' (environment indicator),
and all other experimental design covariates (e.g. replicate, blocks, etc.).
The column names of the data.frame must be identical to the one specified
('genotype', 'check', 'cross', env'). The names of the experimental design
covariates must be the same as the one used in 'exp_des_form'. for more
details see \code{\link{plot_data}}.}

\item{mppData}{An object of class \code{mppData}.}

\item{trait}{\code{Character vector} specifying which traits should be used.}

\item{Q.eff}{\code{Character} expression indicating the assumption concerning
the QTL effects: 1) "cr" for cross-specific; 2) "par" for parental; 3) "anc"
for ancestral; 4) "biall" for a bi-allelic. Default = "cr".}

\item{VCOV}{VCOV \code{Character} expression defining the type of variance
covariance structure used. "CSRT" for within environment
cross-specific residual term, "CS_CSRT" for compound symmetry with within
environment cross-specific residual term. Default = "CS_CSRT".}

\item{exp_des_form}{\code{Character} expression for the random experimental
design effects in asreml-R format. For example,
'env:replicate + env:replicate:block'. The column variables names used in
'exp_des_form' should strictly match the names used in 'plot_data'.}

\item{QTL}{Object of class \code{QTLlist} representing a list of
selected marker positions obtained with the function QTL_select() or
vector of \code{character} marker positions names. Default = NULL.}

\item{glb.only}{\code{Logical} value. if \code{glb.only = TRUE}, the
function only calculate the global R2 for all QTLs. Default = FALSE.}

\item{workspace}{size of workspace for the REML routines measured in double
precision words (groups of 8 bytes). The default is workspace = 8e6.}
}
\value{
Return:

\item{}{}
}
\description{
Compute MPP GxE one stage QTL R2
}
\examples{

# Come later

}
\author{
Vincent Garin
}
