% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppGE_proc.R
\name{mppGE_proc}
\alias{mppGE_proc}
\title{MPP GxE QTL analysis}
\usage{
mppGE_proc(
  pop.name = "MPP",
  trait.name = "trait1",
  mppData,
  trait,
  EnvNames = NULL,
  Q.eff = "cr",
  VCOV = "CS_CSRT",
  plot.gen.eff = FALSE,
  thre.cof = 4,
  win.cof = 50,
  N.cim = 1,
  window = 20,
  thre.QTL = 4,
  win.QTL = 20,
  alpha = 0.01,
  Qeff_est = TRUE,
  text.size = 18,
  parallel = FALSE,
  cluster = NULL,
  workspace = 8e+06,
  verbose = TRUE,
  output.loc = NULL
)
}
\arguments{
\item{pop.name}{\code{Character} name of the studied population.
Default = "MPP".}

\item{trait.name}{\code{Character} name of the studied trait.
Default = "trait1".}

\item{mppData}{An object of class \code{mppData}.}

\item{trait}{\code{Character vector} specifying which traits (environments) should be used.}

\item{EnvNames}{\code{character} expression indicating the environment or trait
labels. By default: Env_1, 2, 3, etc.}

\item{Q.eff}{\code{Character} expression indicating the assumption concerning
the QTL effect: 1) "cr" for cross-specific effects; 2) "par" parental
effects; 3) "anc" for an ancestral effects; 4) "biall" for a bi-allelic
effects. Default = "cr".}

\item{VCOV}{VCOV \code{Character} expression defining the type of variance
covariance structure used. "ID" for identity, "CSRT" for within environment
cross-specific residual terms, "CS_CSRT" for compound symmetry with within
environment cross-specific residual terms. Default = "CS_CSRT".}

\item{plot.gen.eff}{\code{Logical} value. If \code{plot.gen.eff = TRUE},
the function will save the significance of the QTL allelic effects per
cross/parent along the genome. These results can be visualized with the
function \code{\link{plot_genEffects_GE}}. Default value = FALSE.}

\item{thre.cof}{\code{Numeric} value representing the -log10(p-value)
threshold above which a position can be selected as cofactor. Default = 4.}

\item{win.cof}{\code{Numeric} value in centi-Morgan representing the minimum
distance between two selected cofactors. Default = 50 cM.}

\item{N.cim}{\code{Numeric} value specifying the number of times the CIM
analysis is repeated. Default = 1.}

\item{window}{\code{Numeric} distance (cM) on the left and the right of a
cofactor position where it is not included in the model. Default = 20.}

\item{thre.QTL}{\code{Numeric} value representing the -log10(p-value)
threshold above which a position can be selected as QTL. Default = 4.}

\item{win.QTL}{\code{Numeric} value in centi-Morgan representing the minimum
distance between two selected QTLs. Default = 20.}

\item{alpha}{\code{Numeric} value indicating the level of significance for
the backward elimination. Default = 0.01.}

\item{Qeff_est}{\code{Logical} value specifying if the QTL effects of the
final QTL list should be estimated. Default = TRUE.}

\item{text.size}{\code{Numeric} value specifying the size of graph axis text
elements. Default = 18.}

\item{parallel}{\code{Logical} value specifying if the function should be
executed in parallel on multiple cores. To run the function in parallel user must
provide clusters in the \code{cluster} argument. \strong{Parallelization is
only available for 'ID' model}. Default = FALSE.}

\item{cluster}{Cluster object obtained with the function \code{makeCluster()}
from the parallel package. Default = NULL.}

\item{workspace}{Size of workspace for the REML routines measured in double
precision words (groups of 8 bytes). The default is workspace = 8e6.}

\item{verbose}{\code{Logical} value indicating if the steps of mpp_proc should
be printed. It will not affect the printing of the other functions called by
\code{mpp_proc()}, especially the printing of \code{asreml()}. Default = TRUE.}

\item{output.loc}{Path where a folder will be created to save the results.
Default = NULL.}
}
\value{
Return:

List containing the following items:

\item{n.QTL}{Number of detected QTLs.}

\item{cofactors}{\code{Data.frame} with cofactors positions.}

\item{QTL}{\code{Data.frame} with QTL positions.}

\item{R2}{\code{List} containing R squared statistics of the QTL effects.
For details see \code{\link{QTL_R2}} output section.}

Some output files are also saved at the specified location
(\code{output.loc}):

\enumerate{

\item{The SIM and CIM results in a text file (SIM.txt, CIM.txt).}

\item{The list of cofactors (cofactors.txt).}

\item{The list of QTL (QTL.txt).}

\item{The QTL R squared statistics (QTL_R2.txt) (for details see
\code{\link{QTL_R2}}).}

\item{The plot of the CIM profile (QTL_profile.pdf) with dotted vertical
lines representing the cofactors positions. If \code{plot.gen.eff = TRUE},
plot of the genetic effects per cross or parents obtained with
\code{\link{plot_genEffects_GE}}  (gen_eff.pdf) with dashed
lines representing the QTL positions.}

}
}
\description{
QTL detection in MPP characterized in multiple environments.
}
\details{
The procedure is the following:

\enumerate{

\item{Simple interval mapping (SIM) to select cofactor
(\code{\link{mppGE_SIM}}).}

\item{Composite interval mapping (CIM) with selected cofactors
(\code{\link{mppGE_CIM}}).}

\item{Backward elimination on the list of QTLs
(\code{\link{mppGE_back_elim}}).}

\item{Optional estimation of QTLs additive allelic effect
(\code{\link{mppGE_QTL_effects}}).}

}
}
\examples{

\dontrun{

library(asreml)

data(mppData_GE)

# Specify a location where your results will be saved
my.loc <- tempdir()

QTL <- mppGE_proc(pop.name = 'EUNAM', trait.name = 'DMY', mppData = mppData_GE,
                  trait = c('DMY_CIAM', 'DMY_TUM'), EnvNames = c('CIAM', 'TUM'),
                  Q.eff = 'par', plot.gen.eff = TRUE, thre.cof = 3,
                  thre.QTL = 3, verbose = FALSE, output.loc = my.loc)

}

}
\seealso{
\code{\link{mppGE_CIM}},
\code{\link{mppGE_SIM}},
\code{\link{mppGE_back_elim}},
\code{\link{mppGE_QTL_effects}},
\code{\link{plot_genEffects_GE}}
}
\author{
Vincent Garin
}
