% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppGE_QTL_effects.R
\name{mppGE_QTL_effects}
\alias{mppGE_QTL_effects}
\title{MPP GxE QTL genetic effects}
\usage{
mppGE_QTL_effects(mppData, trait, Q.eff = "cr", VCOV = "CS_CSRT",
  QTL = NULL, workspace = 8e+06)
}
\arguments{
\item{mppData}{An object of class \code{mppData}.}

\item{trait}{\code{Character vector} specifying which traits (environments) should be used.}

\item{Q.eff}{\code{Character} expression indicating the assumption concerning
the QTL effects: 1) "cr" for cross-specific; 2) "par" for parental; 3) "anc"
for ancestral; 4) "biall" for a bi-allelic. Default = "cr".}

\item{VCOV}{VCOV \code{Character} expression defining the type of variance
covariance structure used. "ID" for identity, "CSRT" for within environment
cross-specific residual terms, "CS_CSRT" for compound symmetry with within
environment cross-specific residual terms. Default = "CS_CSRT".}

\item{QTL}{Object of class \code{QTLlist} representing a list of
selected marker positions obtained with the function QTL_select() or
a vector of \code{character} marker positions names. Default = NULL.}

\item{workspace}{Size of workspace for the REML routines measured in double
precision words (groups of 8 bytes). The default is workspace = 8e6.}
}
\value{
Return:

\item{Qeff}{\code{List} of \code{data.frame} (one per QTL) containing the
following information:

\enumerate{

\item{QTL genetic effects}
\item{Standard error of the QTL effects.}
\item{Test statistics of the effects (t-test or Wald statistic).}
\item{P-value of the test statistics.}
\item{Significance of the QTL effects.}

}

}
}
\description{
Compute MPP GxE QTL genetic effects
}
\examples{

library(asreml)

data(mppData_GE)

Qpos <- c("PZE.105068880", "PZE.106098900")

Qeff <- mppGE_QTL_effects(mppData = mppData_GE, trait = c('DMY_CIAM', 'DMY_TUM'),
                          Q.eff = 'par', QTL = Qpos)

Qeff

}
\author{
Vincent Garin
}
