% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTL_effect_GE.R
\name{QTL_effects_GE}
\alias{QTL_effects_GE}
\title{MPP GxE QTL genetic effects}
\usage{
QTL_effects_GE(
  mppData,
  trait,
  Q.eff = "par",
  VCOV = "UN",
  QTL = NULL,
  maxIter = 100,
  msMaxIter = 100
)
}
\arguments{
\item{mppData}{An object of class \code{mppData}.}

\item{trait}{\code{Character vector} specifying which traits (environments) should be used.}

\item{Q.eff}{\code{Character} expression indicating the assumption concerning
the QTL effects: 1) "cr" for cross-specific; 2) "par" for parental; 3) "anc"
for ancestral. Default = "par".}

\item{VCOV}{VCOV \code{Character} expression defining the type of variance
covariance structure used. 'CS' for compound symmetry assuming a unique
genetic covariance between environments. 'CSE' for cross-specific within
environment error term. 'CS_CSE' for both compound symmetry plus
cross-specific within environment error term. 'UN' for unstructured
environmental variance covariance structure allowing a specific genotypic
covariance for each pair of environments. Default = 'UN'}

\item{QTL}{Object of class \code{QTLlist} representing a list of
selected marker positions obtained with the function QTL_select() or
a vector of \code{character} marker positions names. Default = NULL.}

\item{maxIter}{maximum number of iterations for the lme optimization algorithm.
Default = 100.}

\item{msMaxIter}{maximum number of iterations for the optimization step inside
the lme optimization. Default = 100.}
}
\value{
Return:

\item{Qeff}{\code{List} of \code{data.frame} (one per QTL) containing the
following information:

\enumerate{

\item{QTL genetic effects}
\item{Standard error of the QTL effects.}
\item{Wald statistics of the effects.}
\item{P-value of the test statistics.}
\item{Significance of the QTL effects.}

}

}
}
\description{
Compute MPP GxE QTL genetic effects.
}
\examples{

data(mppData_GE)

Qpos <- c("PZE.105068880", "PZE.106098900")

Qeff <- QTL_effects_GE(mppData = mppData_GE, trait = c('DMY_CIAM', 'DMY_TUM'),
                          Q.eff = 'par', QTL = Qpos)

Qeff

}
\author{
Vincent Garin
}
