% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTL_effects_GE_1QTL.R
\name{QTL_effects_GE_1QTL}
\alias{QTL_effects_GE_1QTL}
\title{MPP GxE QTL genetic effects for a single QTL}
\usage{
QTL_effects_GE_1QTL(mppData, trait, Q.eff = "cr", VCOV = "CS_CSRT",
  QTL = NULL, Q_sel, workspace = 8e+06)
}
\arguments{
\item{mppData}{An object of class \code{mppData}.}

\item{trait}{\code{Character vector} specifying which traits should be used.}

\item{Q.eff}{\code{Character} expression indicating the assumption concerning
the QTL effects: 1) "cr" for cross-specific; 2) "par" for parental; 3) "anc"
for ancestral; 4) "biall" for a bi-allelic. Default = "cr".}

\item{VCOV}{VCOV \code{Character} expression defining the type of variance
covariance structure used. "ID" for identity, "CSRT" for within environment
cross-specific residual term, "CS_CSRT" for compound symmetry with within
environment cross-specific residual term. Default = "CS_CSRT".}

\item{QTL}{Object of class \code{QTLlist} representing a list of
selected marker positions obtained with the function QTL_select() or
vector of \code{character} marker positions names. Default = NULL.}

\item{Q_sel}{\code{Numeric} indicating the QTL position for which the effect
should be calculated.}

\item{workspace}{size of workspace for the REML routines measured in double
precision words (groups of 8 bytes). The default is workspace = 8e6.}
}
\value{
Return:

\item{Qeff}{\code{List} of \code{data.frame} (one per QTL) containing the
following information:

\enumerate{

\item{QTL genetic effects per cross or parent.}
\item{Standard error of the QTL effects.}
\item{Test statistics of the effects (t-test or Wald statistic).}
\item{P-value of the test statistics.}
\item{Significance of the QTL effects.}

}

}
}
\description{
Compute MPP GxE QTL genetic effects of a single QTL position given a number
of positions set as cofactors.
}
\examples{

# Come later

}
\author{
Vincent Garin
}
