% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTL_effects_oneS_1QTL.R
\name{QTL_effects_oneS_1QTL}
\alias{QTL_effects_oneS_1QTL}
\title{MPP GxE one stage QTL genetic effects for a single QTL}
\usage{
QTL_effects_oneS_1QTL(
  plot_data,
  mppData,
  trait,
  Q.eff = "cr",
  VCOV = "CS_CSRT",
  exp_des_form,
  QTL = NULL,
  Q_sel,
  workspace = 8e+06
)
}
\arguments{
\item{plot_data}{\code{Data.frame} containing the plot data with the following
columns: the trait(s), 'genotype' (genotype indicator), 'check'
(check indicator), 'cross' (cross indicator), 'env' (environment indicator),
and all other experimental design covariates (e.g. replicate, blocks, etc.).
The column names must be ('genotype', 'check', 'cross', env'). The names of
the experimental design covariates must be the same as the one used in
'exp_des_form'. For more details see \code{\link{plot_data}}.}

\item{mppData}{Object of class \code{mppData} contaning the the same
genotype identifiers as the one in \code{plot_data} ('genotype').}

\item{trait}{\code{Character} expression for the trait matching the trait
column in 'plot_data' argument.}

\item{Q.eff}{\code{Character} expression indicating the assumption concerning
the QTL effects: 1) "cr" for cross-specific; 2) "par" for parental; 3) "anc"
for ancestral; 4) "biall" for a bi-allelic. Default = "cr".}

\item{VCOV}{VCOV \code{Character} expression defining the type of variance
covariance structure used: a) "CSRT" for within environment
cross-specific residual terms; b) "CS_CSRT" for compound symmetry with within
environment cross-specific residual terms. Default = "CS_CSRT".}

\item{exp_des_form}{\code{Character} expression for the random experimental
design effects in asreml-R format. For example,
'env:replicate + env:replicate:block'. The variable names used in
'exp_des_form' should strictly match the column names used in 'plot_data'.}

\item{QTL}{Object of class \code{QTLlist} representing a list of
selected marker positions obtained with the function QTL_select() or
a vector of \code{character} marker positions names. Default = NULL.}

\item{Q_sel}{\code{Numeric} indicating the QTL position for which the effect
should be calculated.}

\item{workspace}{Size of workspace for the REML routines measured in double
precision words (groups of 8 bytes). The default is workspace = 8e6.}
}
\value{
Return:

\item{Qeff}{\code{List} of \code{data.frame} (one per QTL) containing the
following information:

\enumerate{

\item{QTL genetic effects per cross or parent.}
\item{Standard error of the QTL effects.}
\item{Test statistics of the effects (Wald statistic).}
\item{P-value of the test statistics.}
\item{Significance of the QTL effects.}

}

}
}
\description{
Compute MPP GxE one stage QTL genetic effects of a single QTL position given
a number of positions set as cofactors.
}
\examples{

# Come later

}
\author{
Vincent Garin
}
