% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTL_R2_GE.R
\name{QTL_R2_GE}
\alias{QTL_R2_GE}
\title{MPP GxE QTL R2}
\usage{
QTL_R2_GE(mppData, trait, Q.eff = "par", QTL = NULL, glb.only = FALSE)
}
\arguments{
\item{mppData}{An object of class \code{mppData}.}

\item{trait}{\code{Character vector} specifying which traits (environments) should be used.}

\item{Q.eff}{\code{Character} expression indicating the assumption concerning
the QTL effects: 1) "cr" for cross-specific; 2) "par" for parental; 3) "anc"
for ancestral; 4) "biall" for a bi-allelic. Default = "par".}

\item{QTL}{Object of class \code{QTLlist} representing a list of
selected marker positions obtained with the function QTL_select() or
a vector of \code{character} marker positions names. Default = NULL.}

\item{glb.only}{\code{Logical} value. If glb.only = TRUE, only the global and
global adjusted R squared will be returned. Default = FALSE.}
}
\value{
Return:

\item{}{}
}
\description{
Compute R2 statistics for MPP GxE QTL using a linear model.
}
\examples{

data(mppData_GE)

Qpos <- c("PZE.105068880", "PZE.106098900")

R2 <- QTL_R2_GE(mppData = mppData_GE, trait = c('DMY_CIAM', 'DMY_TUM'),
                   Q.eff = 'par', QTL = Qpos)

}
\author{
Vincent Garin
}
