% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genEffects_GE.R
\name{plot_genEffects_GE}
\alias{plot_genEffects_GE}
\title{plot of genome wide genetic effect of multi-environment (trait) QTL scan}
\usage{
plot_genEffects_GE(mppData, nEnv, EnvNames, Qprof, Q.eff, QTL = NULL,
  main = "QTL genetic effects plot", text.size = 18)
}
\arguments{
\item{mppData}{An object of class \code{mppData}.}

\item{nEnv}{\code{Numeric} number of environment or trait.}

\item{EnvNames}{\code{character} expression indicating the environment or trait
labels.}

\item{Qprof}{object obtained with function \code{\link{mppGE_SIM}} or
\code{\link{mppGE_CIM}} using argument \code{plot.gen.effect = TRUE}.}

\item{Q.eff}{on of "cr", "par", "anc" or "biall".}

\item{QTL}{Optional argument. Object of class \code{QTLlist} representing a
list of selected position obtained with the function \code{\link{QTL_select}}
or vector of \code{character} marker or in between marker positions names.
These positions will be plotted on the graph. Default = NULL.}

\item{main}{Title of the graph. Default =  "QTL genetic effects plot".}

\item{text.size}{\code{Numeric} value specifying the size of graph axis text
elements. Default = 18.}
}
\description{
plot of genome wide genetic effect of multi-environment (trait) QTL scan
}
\examples{

# Come later

}
\seealso{
\code{\link{mppGE_CIM}}, \code{\link{mppGE_SIM}}
}
\author{
Vincent Garin
}
